+++ UnitTests/GameServer/gameutils/ZoneTest.cs
@@ -30,54 +30,55 @@ namespace DOL.GS.Tests
 	/// </summary>
 	[TestFixture]
 	public class ZoneTest : DOLTestCase
-	{				
+	{
 
 		public ZoneTest()
 		{
 		}
 
-		[Test] public void GetZCoordinateTest()
+		[Test]
+		public void GetZCoordinateTest()
 		{
 
 		}
 
-		[Test] public void GetRandomNPCTest()
+		[Test]
+		public void GetRandomNPCTest()
 		{
 			Zone zone = WorldMgr.GetZone(0);
 			Assert.IsNotNull(zone);
-					
+
 			StartWatch();
 			GameNPC npc = zone.GetRandomNPC(eRealm.None, 5, 7);
-			if (npc!=null) 
+			if (npc != null)
 			{
-				Console.WriteLine("Found NPC from Realm None in "+zone.ZoneRegion.Description+"/"+zone.Description+":"+npc.Name+" level:"+npc.Level);
-				if (npc.Level<5 || npc.Level>7)
+				Console.WriteLine("Found NPC from Realm None in " + zone.ZoneRegion.Description + "/" + zone.Description + ":" + npc.Name + " level:" + npc.Level);
+				if (npc.Level < 5 || npc.Level > 7)
 					Assert.Fail("NPC Level out of defined range");
 
-				if (npc.Realm!= (byte)eRealm.None)
+				if (npc.Realm != eRealm.None)
 					Assert.Fail("NPC wrong Realm");
 			}
-			else 
+			else
 			{
-				Console.WriteLine("nothing found in "+zone.ZoneRegion.Description+"/"+zone.Description);			
+				Console.WriteLine("nothing found in " + zone.ZoneRegion.Description + "/" + zone.Description);
 			}
-			StopWatch();			
+			StopWatch();
 
 			StartWatch();
 			npc = zone.GetRandomNPC(eRealm.Albion);
-			if (npc!=null) 
+			if (npc != null)
 			{
-				Console.WriteLine("Found Albion NPC in "+zone.ZoneRegion.Description+"/"+zone.Description+":"+npc.Name);			
-			
-				if (npc.Realm != (byte)eRealm.Albion)
+				Console.WriteLine("Found Albion NPC in " + zone.ZoneRegion.Description + "/" + zone.Description + ":" + npc.Name);
+
+				if (npc.Realm != eRealm.Albion)
 					Assert.Fail("NPC wrong Realm");
 			}
-			else 
+			else
 			{
-				Console.WriteLine("nothing found in "+zone.ZoneRegion.Description+"/"+zone.Description);
+				Console.WriteLine("nothing found in " + zone.ZoneRegion.Description + "/" + zone.Description);
 			}
 			StopWatch();
-		}			
-
+		}
 	}
 }
